backend/vector_builder/requirements.txt
fastembed>=0.2.0
faiss-cpu>=1.7.4
numpy>=1.21.0
